+++ src/core/exceptions/HttpException.ts
