+++ src/middlewares/error.middleware.ts
@@ -1,5 +1,5 @@
 import { NextFunction, Request, Response } from "express";
-import { HttpException } from "../exceptions/HttpException";
+import { HttpException } from "../core/exceptions/HttpException";
 import Respond from "../services/Respond";
 
 function errorMiddleware(
